Refactor Input System code
Change from cubos/input to cubos/io
Split context, action, manager in multiple files
Fix minor issues (guards, deprecated code, comment typos)
